requirements.txt
django
djangorestframework
django-debug-toolbar
pillow
djangorestframework_simplejwt